refactor(desktop): split heartbeat precheck into dedicated files
拆分 heartbeat.go（889 行）与 heartbeat_test.go（976 行）以回到仓库
800 行文件上限内：precheck 类型/常量/runPrecheck/TestPrecheck 移至新文件
heartbeat_precheck.go，5 个 precheck 测试移至 heartbeat_precheck_test.go。
runPrecheck 的执行逻辑提取为 runPrecheckGate 方法（executeTaskOwned 从
132 行降至 100 行内）。行为等价——仅文件组织调整，无逻辑变更。
baseline.json 仅更新 bridge.ts file-size 4868→4872（新增
HeartbeatTestPrecheck 绑定方法的 4 行契约）。
