+++ tools/repolint/baseline.json
@@ -147,7 +147,7 @@
       "file-size": 180
     },
     "desktop/frontend/src/lib/bridge.ts": {
-      "file-size": 4868
+      "file-size": 4872
     },
     "desktop/frontend/src/lib/crash.ts": {
       "file-size": 225
